delete(useFetchProducts): 기존의 useFetchProducts를 react-query를 사용하는 useFetchProducts2로 대체
